+++ src/rai/rai/utils/ros.py
@@ -1,0 +1,76 @@
+from collections import deque
+from dataclasses import dataclass, field
+from typing import Deque, Dict, List, Optional, Tuple
+
+from langchain_core.language_models import BaseChatModel
+from langchain_core.prompts import ChatPromptTemplate
+
+
+class RosoutBuffer:
+    def __init__(self, llm: BaseChatModel, bufsize: int = 100) -> None:
+        self.bufsize = bufsize
+        self._buffer: Deque[str] = deque()
+        self.template = ChatPromptTemplate.from_messages(
+            [
+                (
+                    "system",
+                    "Shorten the following log keeping its format - for example merge simillar or repeating lines",
+                ),
+                ("human", "{rosout}"),
+            ]
+        )
+        llm = llm
+        self.llm = self.template | llm
+
+    def clear(self):
+        self._buffer.clear()
+
+    def append(self, line: str):
+        self._buffer.append(line)
+        if len(self._buffer) > self.bufsize:
+            self._buffer.popleft()
+
+    def get_raw_logs(self, last_n: int = 30) -> str:
+        return "\n".join(list(self._buffer)[-last_n:])
+
+    def summarize(self):
+        if len(self._buffer) == 0:
+            return "No logs"
+        buffer = self.get_raw_logs()
+        response = self.llm.invoke({"rosout": buffer})
+        return str(response.content)
+
+
+@dataclass
+class NodeDiscovery:
+    topics_and_types: Dict[str, str] = field(default_factory=dict)
+    services_and_types: Dict[str, str] = field(default_factory=dict)
+    actions_and_types: Dict[str, str] = field(default_factory=dict)
+    whitelist: Optional[List[str]] = field(default_factory=list)
+
+    def set(self, topics, services, actions):
+        def to_dict(info: List[Tuple[str, List[str]]]) -> Dict[str, str]:
+            return {k: v[0] for k, v in info}
+
+        self.topics_and_types = to_dict(topics)
+        self.services_and_types = to_dict(services)
+        self.actions_and_types = to_dict(actions)
+        if self.whitelist is not None:
+            self.__filter(self.whitelist)
+
+    def __filter(self, whitelist: List[str]):
+        for d in [
+            self.topics_and_types,
+            self.services_and_types,
+            self.actions_and_types,
+        ]:
+            to_remove = [k for k in d if k not in whitelist]
+            for k in to_remove:
+                d.pop(k)
+
+    def dict(self):
+        return {
+            "topics_and_types": self.topics_and_types,
+            "services_and_types": self.services_and_types,
+            "actions_and_types": self.actions_and_types,
+        }
